requirements.txt
falcon >= 4.2.0
asyncpg == 0.31.0
pyseto >= 1.9.1
aioboto3 >= 15.5.0
uvicorn >= 0.34.0
python-dotenv >= 1.0.0